+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## WuChat
 
+# [1.0.9] - [2018-04-28 15:26] liupei
+1. Sequelize定义模块映射Mysql功能表
+
 # [1.0.8] - [2018-04-27 18:31] liupei
 1. 功能表结构设计
-2. SQL表文件创建，一共九个表(src: server/sequelize/sql/WuChat.sql)
+2. SQL表文件创建，一共九个表(src: server/sequelize/sql/WuChat_2018-04-27.sql)
 
 # [1.0.7] - [2018-04-26 20:45] liupei
 1. 引入Vuex